% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Compute_Power_Expanded}
\alias{Compute_Power_Expanded}
\title{Compute the Power of an association study, at a given sample size，using the idea of expanded dataset}
\usage{
Compute_Power_Expanded(
  parameters,
  n,
  response = "binary",
  covariate = "binary",
  mode = "additive",
  alpha = 0.05,
  seed = 123,
  LargePowerApproxi = F
)
}
\arguments{
\item{parameters}{A list of parameters that contains all the required parameters in the model. If response is "binary", this list needs to contain "prev" which denotes the prevalence of the disease (or case to control ratio for case-control sampling). If response is continuous, the list needs to contain "traitSD" and "traitMean" which represent the standard deviation and mean of the continuous trait.
If covariate is not "none", a parameter "gammaG" needs to be defined to capture the dependence between the SNP and the covariate (through linear regression model if covariate is continuous, and logistic model if covariate is binary). If covariate is "binary", list needs to contains "pE" that defines the frequency of the covariate. If it is continuous, list needs to contain "muE" and "sigmaE" to define
its mean and standard deviation. The MAF is defined as "pG", with HWE assumed to hold.}

\item{n}{An integer number that indicates the sample size.}

\item{response}{A string of either "binary" or "continuous", indicating the type of response/trait variable in the model, by default is "binary"}

\item{covariate}{A string of either "binary", "continuous" or "none" indicating the type of covariate E in the model, by default is "binary".}

\item{mode}{A string of either "additive", "dominant" or "recessive", indicating the genetic mode, by default is "additive".}

\item{alpha}{A numeric value that denotes the significance level used in the study, by default is 0.05.}

\item{seed}{An integer number that indicates the seed used for the simulation to compute the approximate fisher information matrix, by default is 123.}

\item{LargePowerApproxi}{TRUE or FALSE indicates whether to use the large power approximation formula.}
}
\value{
The power that can be ahieved at the given sample size.
}
\description{
Compute the Power of an association study, at a given sample size，using the idea of expanded dataset
}
\examples{
parameters <- list(TraitMean = 0.3, TraitSD = 1, pG = 0.2, betaG = log(1.1), betaE = log(1.1), muE = 0, sigmaE = 3, gammaG = log(2.1))
Compute_Power_Expanded(parameters, n = 1000, "continuous", "continuous")
}
